+++ aichallenge/workspace/src/aichallenge_submit/costmap/booars_costmap_generator/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(multi_layered_costmap_visualizer)
+project(booars_costmap_generator)
 
 if(NOT CMAKE_CXX_STANDARD)
   set(CMAKE_CXX_STANDARD 14)
@@ -14,13 +14,13 @@ endif()
 find_package(ament_cmake_auto REQUIRED)
 ament_auto_find_build_dependencies()
 
-ament_auto_add_library(multi_layered_costmap_visualizer SHARED
-  src/multi_layered_costmap_visualizer.cpp
+ament_auto_add_library(costmap_generator SHARED
+  src/costmap_generator.cpp
 )
 
-rclcpp_components_register_node(multi_layered_costmap_visualizer
-  PLUGIN "multi_layered_costmap_visualizer::MultiLayeredCostmapVisualizer"
-  EXECUTABLE multi_layered_costmap_visualizer_node
+rclcpp_components_register_node(costmap_generator
+  PLUGIN "booars_costmap_generator::CostmapGenerator"
+  EXECUTABLE costmap_generator_node
 )
 
 if(BUILD_TESTING)
